Add splitter and tax-manager
