+++ programs/splitter/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "taxable-sol-token"
+version = "0.1.0"
+description = "Created with Anchor"
+edition = "2021"
+
+[lib]
+crate-type = ["cdylib", "lib"]
+name = "splitter"
+
+[features]
+default = []
+cpi = ["no-entrypoint"]
+no-entrypoint = []
+no-idl = []
+no-log-ix-name = []
+idl-build = ["anchor-lang/idl-build", "anchor-spl/idl-build"]
+
+[dependencies]
+anchor-lang = {version = "0.30.1", features = ["init-if-needed"]}
+anchor-spl = "0.30.1"
