+++ programs/tax-manager/Cargo.toml
@@ -1,12 +1,12 @@
 [package]
-name = "hello-world"
+name = "taxable-sol-token"
 version = "0.1.0"
 description = "Created with Anchor"
 edition = "2021"
 
 [lib]
 crate-type = ["cdylib", "lib"]
-name = "hello_world"
+name = "tax-manager"
 
 [features]
 default = []
@@ -14,7 +14,8 @@ cpi = ["no-entrypoint"]
 no-entrypoint = []
 no-idl = []
 no-log-ix-name = []
-idl-build = ["anchor-lang/idl-build"]
+idl-build = ["anchor-lang/idl-build", "anchor-spl/idl-build"]
 
 [dependencies]
-anchor-lang = "0.30.1"
+anchor-lang = {version = "0.30.1", features = ["init-if-needed"]}
+anchor-spl = "0.30.1"
